Move out incompatibility displaying tooltip to own class
