+++ osu.Game/Overlays/Mods/IncompatibilityDisplayingModButton.cs
@@ -8,11 +8,7 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Cursor;
-using osu.Game.Graphics;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Rulesets;
 using osu.Game.Rulesets.Mods;
-using osu.Game.Screens.Play.HUD;
 using osu.Game.Utils;
 using osuTK;
 
@@ -66,52 +62,5 @@ namespace osu.Game.Overlays.Mods
         }
 
         public override ITooltip<Mod> GetCustomTooltip() => new IncompatibilityDisplayingTooltip();
-
-        private class IncompatibilityDisplayingTooltip : ModButtonTooltip
-        {
-            private readonly OsuSpriteText incompatibleText;
-
-            private readonly Bindable<IReadOnlyList<Mod>> incompatibleMods = new Bindable<IReadOnlyList<Mod>>();
-
-            [Resolved]
-            private Bindable<RulesetInfo> ruleset { get; set; }
-
-            public IncompatibilityDisplayingTooltip()
-            {
-                AddRange(new Drawable[]
-                {
-                    incompatibleText = new OsuSpriteText
-                    {
-                        Margin = new MarginPadding { Top = 5 },
-                        Font = OsuFont.GetFont(weight: FontWeight.Regular),
-                        Text = "Incompatible with:"
-                    },
-                    new ModDisplay
-                    {
-                        Current = incompatibleMods,
-                        ExpansionMode = ExpansionMode.AlwaysExpanded,
-                        Scale = new Vector2(0.7f)
-                    }
-                });
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OsuColour colours)
-            {
-                incompatibleText.Colour = colours.BlueLight;
-            }
-
-            protected override void UpdateDisplay(Mod mod)
-            {
-                base.UpdateDisplay(mod);
-
-                var incompatibleTypes = mod.IncompatibleMods;
-
-                var allMods = ruleset.Value.CreateInstance().AllMods;
-
-                incompatibleMods.Value = allMods.Where(m => m.GetType() != mod.GetType() && incompatibleTypes.Any(t => t.IsInstanceOfType(m))).Select(m => m.CreateInstance()).ToList();
-                incompatibleText.Text = incompatibleMods.Value.Any() ? "Incompatible with:" : "Compatible with all mods";
-            }
-        }
     }
 }
